feat: replace NavbarChatButtons and NavbarPlaceButtons with SortableList fix: [Windows app] Drag and drop does not work for Places #2500
