+++ src/dotnet/UI.Blazor.App/Components/Navbar/ISortableListBackend.cs
@@ -1,0 +1,11 @@
+using ActualChat.UI.Blazor.App.Module;
+
+namespace ActualChat.UI.Blazor.App.Components;
+
+public interface ISortableListBackend
+{
+    public static readonly string JSCreateMethod = $"{BlazorUIAppModule.ImportName}.SortableList.create";
+
+    [JSInvokable]
+    void OnOrderChanged(string[] ids);
+}
