+++ src/dotnet/UI.Blazor.App/exports.ts
@@ -24,8 +24,7 @@ export * from './Components/LeftPanel/logo-button-svg.lit';
 export * from './Components/MarkupEditor/markup-editor';
 export * from './Components/MarkupParts/CodeBlockMarkupView/code-block-markup-view';
 export * from './Components/MentionList/mention-list';
-export * from './Components/Navbar/navbar-place-buttons';
-export * from './Components/Navbar/navbar-chat-buttons';
+export * from './Components/Navbar/sortable-list';
 export * from './Components/Onboarding/phone-verification-cat-svg.lit';
 export * from './Components/SearchPanel/search-panel';
 export * from './Components/Selection/selection-host';
